+++ lizard_portal/templates/portals/valideer_configuraties.js
@@ -23,7 +23,7 @@
 			    columnLines: true,
 			    dataConfig:[
 				{name: 'polder', title: 'Polder',
-				 editable: false, visible: true, width: 250,
+				 editable: false, visible: true, width: 300,
 				 type: 'text'},
 				{name: 'type', title: 'Type', editable: false,
 				 visible: true, width: 100, type: 'text'},
@@ -31,12 +31,15 @@
 				 editable: false, visible: true, width: 150,
 				 type: 'text'},
 				{name: 'date', title: 'Datum', editable: false,
-				 visible: true, width: 125, type: 'text'},
+				 visible: true, width: 175, type: 'text'},
 				{name: 'action', title: 'Actie',
 				 editable: true, visible: true,
 				 width: 100, type: 'combo',
-				 choices: ['Bewaren', 'Valideren']}
+				 choices: [{id:0, name:'Bewaren'}, {id:1, name: 'Valideren'}]}
 			    ],
+			    useAddDeleteButtons: false,
+			    addDeleteIcon: true,
+			    addEditIcon: true,
 			    proxyUrl: '/portal/validate',
 			    proxyParams: {
 			    }